Upgraded to Tmds.DBus v0.11 and regenerated interfaces. All matched and manually pulled over Battery1, IGattService1, GattCharacteristic1, GattDescriptor1, and MediaControl1.  Unit test: Split-up classes
