+++ src/Plugin.BlueZ.Tests/Services/IMyService.cs
@@ -1,0 +1,11 @@
+﻿using System.Threading.Tasks;
+using Tmds.DBus;
+
+namespace Plugin.BlueZ.Tests.Services
+{
+  [DBusInterface("tmds.myservice")]
+  public interface IMyService : IDBusObject
+  {
+    Task<string> SendCommandAsync(string message);
+  }
+}
